Split generic constructor diagnostics tests
